fix(msvc): silence C4018/C4389/C4701/C4177 in own code
C4177 (float_control pragma inside function scope):
  - sorted_double_loop.hpp: replace #pragma float_control with volatile
    double round-trip to force consistent FP precision between count and
    fill passes on MSVC without triggering the pragma scope warning

C4018/C4389 (signed/unsigned mismatch in comparisons):
  - active_space.cpp: cast Eigen vector .size() to size_t
  - state_vector.cpp: cast Eigen occupation vector .size() to size_t
  - autocas_active_space.cpp: cast int64_t num_bins/min_plateau_size to size_t
  - entropy_active_space.cpp: cast Eigen diff.size() to size_t
  - valence_active_space.cpp: cast num_active_electrons to size_t for comparison
  - utils.cpp: cast Eigen .size() and itrn_shell int members to compatible types
  - pipek_mezey.cpp: cast int64_t max_sweeps to size_t
  - vvhv.cpp: cast size_t max_l_ori to int for signed loop variable
  - orbitals.cpp: cast Eigen .rows()/.cols()/.size() to size_t in validations
  - stability.cpp: cast Eigen .rows() to size_t in unrestricted flag checks
  - cholesky_hamiltonian.cpp: cast Eigen .rows() to size_t in validations
  - basis_set.cpp: cast Eigen nuclear_charges.size() to size_t in loop bound

C4701 (potentially uninitialized local variable):
  - scf.cpp: default-initialize BasisSetType before if/else-if chain
  - cholesky_hamiltonian.cpp: zero-initialize s1_max, s2_max, q_shell_pair_max
  - gauxc_impl.cpp: value-initialize prim_array exponents{} and coefficients{}
  - line_search.h: initialize step_hi=0 and fx_hi=fx0 alongside step_lo/fx_lo

+++ cpp/src/qdk/chemistry/algorithms/microsoft/active_space/autocas_active_space.cpp
@@ -46,15 +46,15 @@ std::shared_ptr<data::Wavefunction> AutocasActiveSpaceSelector::_run_impl(
   // create discrete bins
   const double bin_width = 1.0 / num_bins;
   QDK_LOGGER().debug("Bin width: {:.6f}", bin_width);
-  std::vector<double> bins(num_bins);
-  for (size_t i = 0; i < num_bins; ++i) {
+  std::vector<double> bins(static_cast<size_t>(num_bins));
+  for (size_t i = 0; i < static_cast<size_t>(num_bins); ++i) {
     bins[i] = i * bin_width;
   }
 
   // Count orbitals above each bin
-  std::vector<size_t> orbitals_above_bin(num_bins);
+  std::vector<size_t> orbitals_above_bin(static_cast<size_t>(num_bins));
   size_t entropy_idx = sorted_entropies.size();
-  for (size_t i = 0; i < num_bins; ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(num_bins); ++i) {
     double bin_threshold = bins[i];
     // Move backwards while entropies are <= threshold
     while (entropy_idx > 0 &&
@@ -66,7 +66,7 @@ std::shared_ptr<data::Wavefunction> AutocasActiveSpaceSelector::_run_impl(
   }
   QDK_LOGGER().debug("Orbitals above each bin threshold:");
   QDK_LOGGER().debug("    Bin   Bin Threshold    #Orbitals Above");
-  for (size_t i = 0; i < num_bins; ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(num_bins); ++i) {
     QDK_LOGGER().debug("  {:>5}    {:>12.6f}    {:>15}", i, bins[i],
                        orbitals_above_bin[i]);
   }
@@ -81,7 +81,7 @@ std::shared_ptr<data::Wavefunction> AutocasActiveSpaceSelector::_run_impl(
     if (orbitals_above_bin[i] != orbitals_above_bin[i - 1]) {
       // end of plateau
       const size_t plateau_size = i - current_bin;
-      if (plateau_size >= min_plateau_size) {
+      if (plateau_size >= static_cast<size_t>(min_plateau_size)) {
         plateaus.emplace_back(current_bin, i - 1);
       }
       current_bin = i;
@@ -89,7 +89,7 @@ std::shared_ptr<data::Wavefunction> AutocasActiveSpaceSelector::_run_impl(
   }
   // check last plateau
   const size_t plateau_size = bins.size() - current_bin;
-  if (plateau_size >= min_plateau_size) {
+  if (plateau_size >= static_cast<size_t>(min_plateau_size)) {
     plateaus.emplace_back(current_bin, bins.size() - 1);
   }
   // Logging
