+++ cpp/src/qdk/chemistry/algorithms/microsoft/localization/vvhv.cpp
@@ -800,7 +800,7 @@ Eigen::MatrixXd VVHVLocalization::localize_hard_virtuals(
     // First collect all atomic orbitals on atom A for both basis sets
     std::vector<int> bf_list_ori;
     std::vector<int> bf_list_min;
-    for (auto l = 0; l <= max_l_ori; ++l) {
+    for (auto l = 0; l <= static_cast<int>(max_l_ori); ++l) {
       auto& bf_l_ori = al_to_bf_ori[atom_a][l];
       auto& bf_l_min = al_to_bf_min[atom_a][l];
       bf_list_ori.insert(bf_list_ori.end(), bf_l_ori.begin(), bf_l_ori.end());
@@ -838,7 +838,7 @@ Eigen::MatrixXd VVHVLocalization::localize_hard_virtuals(
         Eigen::MatrixXd::Zero(num_atomic_orbitals_ori, nhv_a);
     int proto_hv_idx = 0;
 
-    for (auto l = 0; l <= max_l_ori; ++l) {
+    for (auto l = 0; l <= static_cast<int>(max_l_ori); ++l) {
       auto& bf_al_ori = al_to_bf_ori[atom_a][l];
       if (bf_al_ori.size() == 0)
         continue;  // no atomic orbitals with this angular momentum on this atom
